Converting NSIntegers to longs

+++ 8tracks4mac/UCRemoteCall.h
@@ -6,7 +6,7 @@
 @interface UCRemoteCall : NSObject
 
 -(NSString *)mixes;
--(UCMix *)detailsOfMix:(NSInteger)mix;
+-(UCMix *)detailsOfMix:(long)mix;
 -(UCToken *)playToken;
 -(NSString *)findMixesWithTag:(NSString *)tag1 andTag:(NSString *)tag2;
 -(UCTrack *)trackFromMix:(UCMix *)mix andToken:(UCToken *)token;
